requirements.txt
pandas
numpy
streamlit
matplotlib
datetime
yfinance
python-bcb